At The Island - Version 1
At The Island is a VR experience showcasing VR Ed-Tech companies with a setting on a dreamy island. The main purpose of this work is to let know the people about how these companies are leveraging VR for providing high-quality education. The targeted audience is anyone who is enthusiastic about VR and especially VR in Education.